requirements.txt
schedule
feedparser==6.0.10
requests==2.31.0
openai==0.28.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
slack-sdk==3.23.0
python-dotenv==1.0.0